requirements.txt
streamlit
python-dotenv
PyMuPDF
langchain
langchain-community
langchain-groq
